restructuring and visual adjustments
