--- app/app-root/app-root.ios.css
@@ -1,1 +1,0 @@
-.sidedrawer.sidedrawer-left{background-color:#3A53FF}.sidedrawer.sidedrawer-left .sidedrawer-header-image{color:#F8F8F8;height:60;width:60;font-size:60;padding:0;margin-bottom:5;margin-top:20}.sidedrawer.sidedrawer-left .footnote{color:rgba(255,255,255,0.5)}.sidedrawer .sidedrawer-header{background-color:#3A53FF}.sidedrawer .sidedrawer-header .sidedrawer-header-brand{color:#fff}.sidedrawer .sidedrawer-content{background-color:#fff}.sidedrawer .sidedrawer-list-item{padding-left:15}.sidedrawer .sidedrawer-list-item Label{vertical-align:center;color:#022734}.sidedrawer .sidedrawer-list-item .fa{width:20;margin-right:10}.sidedrawer .sidedrawer-list-item.selected{background-color:#F8F8F8}.sidedrawer .sidedrawer-list-item.selected Label{color:#3A53FF}
