+++ app/app.ts
@@ -6,7 +6,7 @@ purpose of the file is to pass control to the app’s first module.
 import * as app from "tns-core-modules/application";
 import "./bundle-config";
 
-app.run({ moduleName: "app-root/app-root" });
+app.run({ moduleName: "views/app-root/app-root" });
 
 /*
 Do not place any code after the application has been started as it will not
