+++ app/bundle-config.ts
@@ -1,5 +1,4 @@
 if ((global).TNS_WEBPACK) {
     // Register custom modules
-    global.registerModule("nativescript-ui-sidedrawer",
-        () => require("../node_modules/nativescript-ui-sidedrawer"));
+    global.registerModule("nativescript-ui-sidedrawer", () => require("../node_modules/nativescript-ui-sidedrawer"));
 }
