--- app/home/home-page.xml
@@ -1,66 +1,0 @@
-<Page class="page" navigatingTo="onNavigatingTo" 
-    xmlns="http://schemas.nativescript.org/tns.xsd">
-
-    <ActionBar class="action-bar">
-        <!-- 
-        Use the NavigationButton as a side-drawer button in Android
-        because ActionItems are shown on the right side of the ActionBar
-        -->
-        <NavigationButton ios:visibility="collapsed" icon="res://menu" tap="onDrawerButtonTap"></NavigationButton>
-        <!-- 
-        Use the ActionItem for IOS with position set to left. Using the
-        NavigationButton as a side-drawer button in iOS is not possible,
-        because its function is to always navigate back in the application.
-        -->
-        <ActionItem icon="res://navigation/menu" android:visibility="collapsed" tap="onDrawerButtonTap" ios.position="left">
-        </ActionItem>
-        <Label class="action-bar-title" text="Assetti Notes"></Label>
-
-    </ActionBar>
-
-    <TabView id="tabViewContainer">
-        <TabViewItem title="Proximity">
-            <ListView items="{{ myTitles }}" itemTap="onItemTap" loaded="{{ onListViewLoaded }}" separatorColor="gray" rowHeight="auto" class="list-group" id="listView" row="2">
-                <ListView.itemTemplate>
-                    <!-- The item template can only have a single root view container (e.g. GriLayout, StackLayout, etc.) -->
-                    <GridLayout rows="*,*,*" columns="*,auto" class="list-group-item note-list-item">
-                        <Label row="0" col="0" colSpan="2" text="{{ title || 'Blablabla...' }}" textWrap="true" class="title"></Label>
-                        <Label row="1" col="0" text="Note preview text here..."></Label>
-                        <StackLayout row="0" col="1" orientation="horizontal">
-                            <Label text="1"></Label>
-                            <Label text="o" class="icon"></Label>
-                            <Label text="1"></Label>
-                            <Label text="o" class="icon"></Label>
-                        </StackLayout>
-                        <Label row="2" col="0" text="2018-01-01"></Label>
-                        <Label row="2" col="1" text="{{ author || 'Anonymous' }}"></Label>
-                    </GridLayout>
-                </ListView.itemTemplate>
-            </ListView>
-        </TabViewItem>
-        <TabViewItem title="Urgent">
-            <StackLayout>
-                <FlexboxLayout flexDirection="column" backgroundColor="lightgray">
-                    <Label text="Label 2" backgroundColor="red"/>
-                    <Label text="Label 2" backgroundColor="green"/>
-                    <Label text="Label 3" backgroundColor="blue"/>
-                    <Label text="Label 4" backgroundColor="yellow"/>
-                </FlexboxLayout>∏
-            </StackLayout>
-        </TabViewItem>
-        <TabViewItem title="My Notes">
-            <StackLayout>
-                <FlexboxLayout flexDirection="column" backgroundColor="lightgray">
-                    <Label text="Label 2" backgroundColor="red"/>
-                    <Label text="Label 2" backgroundColor="green"/>
-                    <Label text="Label 3" backgroundColor="blue"/>
-                    <Label text="Label 4" backgroundColor="yellow"/>
-                </FlexboxLayout>∏
-            </StackLayout>
-        </TabViewItem>
-    </TabView>
-
-
-
-
-</Page>
