--- app/home/home-view-model.ts
@@ -1,11 +1,0 @@
-import { Observable } from "tns-core-modules/data/observable";
-
-import { SelectedPageService } from "../shared/selected-page-service";
-
-export class HomeViewModel extends Observable {
-    constructor() {
-        super();
-
-        SelectedPageService.getInstance().updateSelectedPage("Home");
-    }
-}
