+++ app/views/app-root/_app-root-common.scss
@@ -1,5 +1,5 @@
 // Custom common variables
-@import '../app-variables';
+@import '~/app-variables';
 
 // Drawer navigation custom styles
 $sidedrawer-header-image-size: 60;
