+++ app/views/app-root/app-root.android.scss
@@ -1,3 +1,3 @@
-@import 'app-root-common';
+@import './app-root-common';
 
 // Place any CSS rules you want to apply only on Android here
