+++ app/views/app-root/app-root.ios.scss
