+++ app/views/newnote/newnote-page.css
