+++ app/views/newnote/newnote-page.ts
