+++ app/views/newnote/newnote-page.xml
@@ -1,28 +1,10 @@
-<Page
-    class="page"
-    navigatingTo="onNavigatingTo"  
-    xmlns="http://schemas.nativescript.org/tns.xsd"
-    loaded="mainload">
-    
+<Page class="page" navigatingTo="onNavigatingTo" xmlns="http://schemas.nativescript.org/tns.xsd" loaded="mainload">
 
-    <ActionBar class="action-bar">
-        <!-- 
-        Use the NavigationButton as a side-drawer button in Android
-        because ActionItems are shown on the right side of the ActionBar
-        -->
-        <NavigationButton ios:visibility="collapsed" icon="res://menu" tap="onDrawerButtonTap"></NavigationButton>
-        <!-- 
-        Use the ActionItem for IOS with position set to left. Using the
-        NavigationButton as a side-drawer button in iOS is not possible,
-        because its function is to always navigate back in the application.
-        -->
-        <ActionItem icon="res://navigation/menu" 
-            android:visibility="collapsed" 
-            tap="onDrawerButtonTap"
-            ios.position="left">
-        </ActionItem>
-        <Label class="action-bar-title" text="New Note"></Label>
+    <ActionBar title="New Note" class="action-bar">
+        <NavigationButton ios:visibility="collapsed" icon="res://menu" tap="onDrawerButtonTap" />
+        <ActionItem icon="res://navigation/menu" android:visibility="collapsed" tap="onDrawerButtonTap" ios.position="left" />
     </ActionBar>
+
     <GridLayout rows="*,auto">
         <FlexboxLayout id="a" class="form" row="0" flexDirection="column">
             <TextField id="notetitle" hint="Title" autocapitalizationType="sentences" style="font-weight: bold; color: gray" class="input input-border" />
@@ -37,4 +19,5 @@
             <Button col="2" text="NEXT" class="btn" tap="changeTab"/>
         </GridLayout>
     </GridLayout>   
+
 </Page>
