+++ app/views/notes-list/notes-list-page.xml
@@ -1,0 +1,49 @@
+<Page class="page" navigatingTo="onNavigatingTo" xmlns="http://schemas.nativescript.org/tns.xsd">
+
+    <ActionBar title="Notes" class="action-bar">
+        <NavigationButton ios:visibility="collapsed" icon="res://menu" tap="onDrawerButtonTap" />
+        <ActionItem icon="res://navigation/menu" android:visibility="collapsed" tap="onDrawerButtonTap" ios.position="left" />
+    </ActionBar>
+
+    <ListView items="{{ notes }}" itemTap="onItemTap" loaded="{{ onListViewLoaded }}" separatorColor="gray" rowHeight="auto" id="listView">
+        <ListView.itemTemplate>
+            <!-- The item template can only have a single root view container (e.g. GriLayout, StackLayout, etc.) -->
+            <DockLayout stretchLastChild="false" class="list-view-item p-15">
+                <Label dock="left" text="[image]" />
+                <StackLayout orientation="vertical" class="p-l-15 p-r-15">
+                    <Label class="font-weight-bold" text="{{ title || 'Loading...'}}" textWrap="true" />
+                    <Label text="{{ author || 'Anonymous' }}" textWrap="true" />
+                </StackLayout>
+                <Label dock="right" text="[??]" />
+            </DockLayout >
+        </ListView.itemTemplate>
+    </ListView>
+
+    <!--
+    tabview commented out here for the time being (not needed)
+    <TabView id="tabViewContainer">
+        <TabViewItem title="Proximity">
+        </TabViewItem>
+        <TabViewItem title="Urgent">
+            <StackLayout>
+                <FlexboxLayout flexDirection="column" backgroundColor="lightgray">
+                    <Label text="Label 2" backgroundColor="red"/>
+                    <Label text="Label 2" backgroundColor="green"/>
+                    <Label text="Label 3" backgroundColor="blue"/>
+                    <Label text="Label 4" backgroundColor="yellow"/>
+                </FlexboxLayout>∏
+            </StackLayout>
+        </TabViewItem>
+        <TabViewItem title="My Notes">
+            <StackLayout>
+                <FlexboxLayout flexDirection="column" backgroundColor="lightgray">
+                    <Label text="Label 2" backgroundColor="red"/>
+                    <Label text="Label 2" backgroundColor="green"/>
+                    <Label text="Label 3" backgroundColor="blue"/>
+                    <Label text="Label 4" backgroundColor="yellow"/>
+                </FlexboxLayout>∏
+            </StackLayout>
+        </TabViewItem>
+    </TabView>
+    -->
+</Page>
