+++ app/views/search/search-page.xml
@@ -1,0 +1,13 @@
+<Page navigatingTo="onNavigatingTo" xmlns="http://schemas.nativescript.org/tns.xsd">
+
+    <ActionBar title="Search" class="action-bar">
+        <NavigationButton ios:visibility="collapsed" icon="res://menu" tap="onDrawerButtonTap" />
+        <ActionItem icon="res://navigation/menu" android:visibility="collapsed" tap="onDrawerButtonTap" ios.position="left" />
+    </ActionBar>
+
+    <GridLayout class="page-content">
+        <Label class="page-icon fa" text="&#xf002;" />
+        <Label class="page-placeholder" text="<!-- Page content goes here -->" />
+    </GridLayout>
+
+</Page>
