+++ app/views/search/search-view-model.ts
@@ -1,6 +1,6 @@
 import { Observable } from "tns-core-modules/data/observable";
 
-import { SelectedPageService } from "../shared/selected-page-service";
+import { SelectedPageService } from "~/shared/selected-page-service";
 
 export class SearchViewModel extends Observable {
     constructor() {
