+++ package.json
@@ -15,6 +15,7 @@
     "lint": "tslint \"app/**/*.ts\""
   },
   "dependencies": {
+    "nativescript": "^5.2.3",
     "nativescript-camera": "^4.2.0",
     "nativescript-theme-core": "~1.0.4",
     "nativescript-ui-chart": "^3.11.1",
